adding compiler module and compiler passes.
